requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
websockets==12.0
aiohttp==3.9.1
beautifulsoup4==4.12.2
python-dotenv==1.0.0
backoff==2.2.1